app/requirements.txt
gpiozero>=2.0
paho-mqtt>=1.6,<2.0
flask>=3.0
flask-socketio>=5.3
gevent>=23.0
requests>=2.31
